# api/swagger.yaml
openapi: 3.0.0
info:
  title: Personal Finance Manager API
  description: API documentation for the Personal Finance Manager application
  version: 1.0.0
  contact:
    name: Backend Team
servers:
  - url: http://localhost:8080
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [income, expense, transfer]
        icon:
          type: string
        color:
          type: string
        parent_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Account:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        plaid_account_id:
          type: string
          nullable: true
        name:
          type: string
        type:
          type: string
        balance:
          type: number
          format: double
        currency:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        category_id:
          type: string
          format: uuid
          nullable: true
        amount:
          type: number
          format: double
        description:
          type: string
        date:
          type: string
          format: date-time
        type:
          type: string
          enum: [income, expense, transfer]
        status:
          type: string
          enum: [pending, completed, cancelled]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Budget:
      type: object
      properties:
        id:
          type: string
          format: uuid
        category_id:
          type: string
          format: uuid
        amount:
          type: number
          format: double
        spent:
          type: number
          format: double
        period:
          type: string
          enum: [monthly, quarterly, yearly]
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
paths:
...
  /api/categories:
    get:
      summary: Get all categories
      tags: [Categories]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
    post:
      summary: Create a new category
      tags: [Categories]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - type
              properties:
                name:
                  type: string
                type:
                  type: string
                  enum: [income, expense, transfer]
                icon:
                  type: string
                color:
                  type: string
                parent_id:
                  type: string
                  format: uuid
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
  /api/accounts:
    get:
      summary: Get user accounts
      tags: [Accounts]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
    post:
      summary: Create a new account
      tags: [Accounts]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - type
                - currency
              properties:
                name:
                  type: string
                type:
                  type: string
                currency:
                  type: string
                plaid_account_id:
                  type: string
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
  /api/transactions:
    get:
      summary: Get user transactions
      tags: [Transactions]
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: start_date
          schema:
            type: string
            format: date
        - in: query
          name: end_date
          schema:
            type: string
            format: date
        - in: query
          name: category_id
          schema:
            type: string
            format: uuid
        - in: query
          name: account_id
          schema:
            type: string
            format: uuid
        - in: query
          name: type
          schema:
            type: string
            enum: [income, expense, transfer]
      responses:
        '200':
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
    post:
      summary: Create a new transaction
      tags: [Transactions]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - account_id
                - amount
                - description
                - date
                - type
              properties:
                account_id:
                  type: string
                  format: uuid
                category_id:
                  type: string
                  format: uuid
                amount:
                  type: number
                  format: double
                description:
                  type: string
                date:
                  type: string
                  format: date-time
                type:
                  type: string
                  enum: [income, expense, transfer]
      responses:
        '201':
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
  /api/budgets:
    get:
      summary: Get user budgets
      tags: [Budgets]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of budgets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Budget'
    post:
      summary: Create a new budget
      tags: [Budgets]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - category_id
                - amount
                - period
              properties:
                category_id:
                  type: string
                  format: uuid
                amount:
                  type: number
                  format: double
                period:
                  type: string
                  enum: [monthly, quarterly, yearly]
                start_date:
                  type: string
                  format: date
                end_date:
                  type: string
                  format: date
      responses:
        '201':
          description: Budget created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Budget'
...